Move `PaginationInfo` to its own module
